return long it for get_number_of_available_processors

+++ src/realtime_helpers.cpp
@@ -184,12 +184,12 @@ std::pair<bool, std::string> set_current_thread_affinity(int core)
   return set_thread_affinity(0, core);
 }
 
-int get_number_of_available_processors()
+int64_t get_number_of_available_processors()
 {
 #ifdef _WIN32
   SYSTEM_INFO sysinfo;
   GetSystemInfo(&sysinfo);
-  return sysinfo.dwNumberOfProcessors;
+  return static_cast<int64_t>(sysinfo.dwNumberOfProcessors);
 #else
   return sysconf(_SC_NPROCESSORS_ONLN);
 #endif
